Project 9 with challenges 1 and 2
Create two arrow shapes: one composed of a rectangle and a triangle, and one drawn as a continuous path of lines. Both are controlled by the shaft length ratio (proportional to the bounding rectangle width) and the shaft and tip ratios (proportional to the bounding rectangle height). Both arrows are displayed inside a red rectangle that is drawn over the bounding rectangle and both are filled and outlined (to show the path differences). The main view adds three sliders to modify the ratios and each supports a tap gesture that chooses three random ratios (reflected in the sliders).