Project 9 with challenge 3
Did some more work on the Arrow shape to make it an Insettable shape. Create the color cycling rectangle view and, along the way, found out about the concept of the UnitPoint, its coordinate, and its relationship to .top, .bottom, etc. Put the view in a TabView along with the updated animated arrows because why wouldn't I? :-)